cleanup + restructure
