+++ api/health.go
@@ -1,0 +1,8 @@
+package api
+
+import "net/http"
+
+// healthz is a liveness probe.
+func healthz(w http.ResponseWriter, _ *http.Request) {
+	w.WriteHeader(http.StatusOK)
+}
